Split benchmarks and svgtools and make new library for valudation utils
