+++ tests/include/detray/test/cpu/detector_scan.hpp
@@ -19,10 +19,10 @@
 #include "detray/test/common/detail/whiteboard.hpp"
 #include "detray/test/common/detector_scan_config.hpp"
 #include "detray/test/common/fixture_base.hpp"
-#include "detray/test/utils/detector_scan_utils.hpp"
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/simulation/event_generator/track_generators.hpp"
 #include "detray/test/utils/types.hpp"
+#include "detray/test/validation/detector_scan_utils.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
 
 // System include(s)
 #include <iostream>
