+++ tests/include/detray/test/cpu/material_validation.hpp
@@ -15,7 +15,7 @@
 // Detray test include(s)
 #include "detray/test/common/fixture_base.hpp"
 #include "detray/test/common/material_validation_config.hpp"
-#include "detray/test/utils/material_validation_utils.hpp"
+#include "detray/test/validation/material_validation_utils.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/host_memory_resource.hpp>
