+++ tests/include/detray/test/device/cuda/material_validation.hpp
@@ -15,7 +15,7 @@
 #include "detray/test/common/fixture_base.hpp"
 #include "detray/test/common/material_validation_config.hpp"
 #include "detray/test/cpu/material_validation.hpp"
-#include "detray/test/utils/material_validation_utils.hpp"
+#include "detray/test/validation/material_validation_utils.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/cuda/device_memory_resource.hpp>
