+++ tests/include/detray/test/device/cuda/navigation_validation.hpp
@@ -18,11 +18,11 @@
 // Detray test include(s)
 #include "detray/test/common/fixture_base.hpp"
 #include "detray/test/common/navigation_validation_config.hpp"
-#include "detray/test/utils/detector_scan_utils.hpp"
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/inspectors.hpp"
-#include "detray/test/utils/material_validation_utils.hpp"
-#include "detray/test/utils/navigation_validation_utils.hpp"
+#include "detray/test/validation/detector_scan_utils.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
+#include "detray/test/validation/material_validation_utils.hpp"
+#include "detray/test/validation/navigation_validation_utils.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/cuda/device_memory_resource.hpp>
