+++ tests/include/detray/test/validation/detector_scanner.hpp
