+++ tests/include/detray/test/validation/svg_display.hpp
