+++ tests/tools/python/file_checker.py
