+++ tests/tools/python/impl/plot_ray_scan.py
