+++ tests/tools/python/json_schema/geometry.py
