+++ tests/tools/python/json_schema/homogeneous_material.py
