+++ tests/tools/python/json_schema/material_maps.py
