+++ tests/tools/python/json_schema/surface_grids.py
