+++ tests/tools/python/navigation_validation.py
