+++ tests/tools/python/options/detector_io_options.py
