+++ tests/tools/root/jacobian_histogram.C
