+++ tests/unit_tests/svgtools/intersections.cpp
@@ -14,10 +14,10 @@
 #include "detray/plugins/svgtools/writer.hpp"
 
 // Detray test include(s)
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/detectors/build_toy_detector.hpp"
 #include "detray/test/utils/simulation/event_generator/track_generators.hpp"
-#include "detray/test/utils/svg_display.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
+#include "detray/test/validation/svg_display.hpp"
 #include "detray/tracks/tracks.hpp"
 
 // Vecmem include(s)
