+++ tests/unit_tests/svgtools/trajectories.cpp
@@ -16,9 +16,9 @@
 #include "detray/plugins/svgtools/writer.hpp"
 
 // Detray test include(s)
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/detectors/build_toy_detector.hpp"
-#include "detray/test/utils/svg_display.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
+#include "detray/test/validation/svg_display.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/host_memory_resource.hpp>
