+++ tutorials/src/cpu/CMakeLists.txt
@@ -29,5 +29,5 @@ foreach(file ${source_files})
    detray_add_tutorial( ${file_name}
       ${file}
       LINK_LIBRARIES detray::core_array covfie::core vecmem::core detray::io
-                     detray::tutorial detray::test_utils )
+                     detray::tutorial detray::validation_utils )
 endforeach()
