+++ tutorials/src/cpu/detector/detector_ray_scan.cpp
@@ -8,11 +8,11 @@
 // Project include(s)
 #include "detray/navigation/detail/ray.hpp"
 #include "detray/navigation/volume_graph.hpp"
-#include "detray/test/utils/detector_scan_utils.hpp"
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/detectors/build_toy_detector.hpp"
 #include "detray/test/utils/hash_tree.hpp"
 #include "detray/test/utils/simulation/event_generator/track_generators.hpp"
+#include "detray/test/validation/detector_scan_utils.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
 
 // Example linear algebra plugin: std::array
 #include "detray/tutorial/types.hpp"
