+++ tutorials/src/cpu/propagation/navigation_inspection.cpp
@@ -12,10 +12,10 @@
 #include "detray/propagator/actor_chain.hpp"
 #include "detray/propagator/line_stepper.hpp"
 #include "detray/propagator/propagator.hpp"
-#include "detray/test/utils/detector_scanner.hpp"
 #include "detray/test/utils/detectors/build_toy_detector.hpp"
 #include "detray/test/utils/inspectors.hpp"
 #include "detray/test/utils/simulation/event_generator/track_generators.hpp"
+#include "detray/test/validation/detector_scanner.hpp"
 #include "detray/tracks/tracks.hpp"
 
 // Example linear algebra plugin: std::array
